Getting new projects for integration tests
