+++ test/Microsoft.SqlTools.SqlCore.IntegrationTests/Microsoft.SqlTools.SqlCore.IntegrationTests.csproj
