+++ test/Microsoft.SqlTools.SqlCore.UnitTests/Scripting/ScriptingObjectMatcherTests.cs
@@ -11,7 +11,7 @@ using Microsoft.SqlTools.SqlCore.Scripting;
 using Microsoft.SqlTools.SqlCore.Scripting.Contracts;
 using NUnit.Framework;
 
-namespace Microsoft.SqlTools.ServiceLayer.UnitTests.Scripting
+namespace Microsoft.SqlTools.SqlCore.UnitTests.Scripting
 {
     public class ScriptingMatcherTests
     {
